backend/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
channels==4.2.2
Django==5.2.1
django-allauth==65.8.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-polymorphic==3.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google-auth==2.40.1
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
rsa==4.9.1
setuptools==80.7.1
sqlparse==0.5.3
